Move identity helpers to own module and drop default_registry
MigrationError and the _identity/_has_identity helpers move to syft_migration/identity.py, breaking the registry<->schema import cycle so registry.py imports the schema classes at the top instead of inside methods. default_registry is removed: a concrete MigratableObject subclass must now receive registry= (or inherit it); defining one without a registry raises MigrationError.
